Made transaction module its own package.
