+++ iota/transaction/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+from __future__ import absolute_import, division, print_function, \
+  unicode_literals
+
+# Import symbols for backwards-compatibility with PyOTA 1.1.x.
+from .read import *
+from .types import *
+from .utils import *
+from .write import *
